Launcher3: Let's change at a glance fonts
This will allows the user to change between fonts on launcher. Pretty simple, it switches
from system to Google Sans. Kind of simple, but for adding new fonts it could be
easy to do by copypasting the googlesans line but with your desired fonts on assets.

bootleggers note: This is just for future tweaks, that's why it's a listpreference
you can easily do a boolean switch but meh

Change-Id: Ia2a2c449d4d3f0f72cbffcbef12c451d07e6a521
Signed-off-by: Mrinal Ghosh <mg712702@gmail.com>