Fix 2 bugs involving split staging task translation
This patch fixes two bugs related to the way tiles slide in when a split is staged. Previously, translations were not being completely reset, so on repeated splits, the thumbnails would retain their old translation setting and cause awkward movements to occur during the slide-in animation.

Fixed by adding setTaskViewsPrimarySplitTranslation() and setTaskViewsSecondarySplitTranslation() in resetFromSplitSelectionState().

Fixes: 249154347
Fixes: 249824716
Test: Manual on grid and non-grid Overview, in a variety of conditions
Change-Id: I2d9aa621a17f178ff8216740d76429d2ac484b39